+++ crates/sncast/tests/e2e/invoke.rs
@@ -12,10 +12,7 @@ use indoc::indoc;
 use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
 use snapbox::cmd::{Command, cargo_bin};
 use sncast::AccountType;
-use sncast::helpers::constants::{
-    // ARGENT_CLASS_HASH, BRAAVOS_CLASS_HASH,
-    OZ_CLASS_HASH,
-};
+use sncast::helpers::constants::{ARGENT_CLASS_HASH, OZ_CLASS_HASH};
 use sncast::helpers::fee::FeeArgs;
 use starknet::core::types::TransactionReceipt::Invoke;
 use starknet_types_core::felt::Felt;
@@ -63,8 +60,8 @@ async fn test_happy_case_human_readable() {
 
 #[test_case(DEVNET_OZ_CLASS_HASH_CAIRO_0.parse().unwrap(), AccountType::OpenZeppelin; "cairo_0_class_hash")]
 #[test_case(OZ_CLASS_HASH, AccountType::OpenZeppelin; "cairo_1_class_hash")]
+#[test_case(ARGENT_CLASS_HASH, AccountType::Argent; "argent_class_hash")]
 // TODO(#3089)
-// #[test_case(ARGENT_CLASS_HASH, AccountType::Argent; "argent_class_hash")]
 // #[test_case(BRAAVOS_CLASS_HASH, AccountType::Braavos; "braavos_class_hash")]
 #[tokio::test]
 async fn test_happy_case(class_hash: Felt, account_type: AccountType) {